Rename 'components/TimeBarChart.vue' to 'components/index/_shared/charts/BarChart.vue'
