+++ components/index/_shared/charts/BarChart.vue
